refactor major objects
Storing our major objects in a list allows us to have code iterate
over that list, rather than a huge if/else block, which has to know
about all the possible items in the list, and how to get to the next
one.

This is cleaner, simpler, and easier to change (say, if we ever wanted
to add Pluto back in, or Ceres, or any other solar system object).